src/producer/requirements.txt
kafka-python==2.0.2
Flask==2.1.2
